backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "snaky-social-hub-backend"
version = "0.1.0"
description = "FastAPI backend for Snaky Social Hub"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-multipart==0.0.6",
    "email-validator>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "httpx==0.25.2",
    "pytest-asyncio==0.21.1",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "httpx>=0.25.2",
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
]
